Refactores addProblems page
Created a forum handler and moved addProblems functions to it. Also created a test cases class to handle the adding and removing of test case functionality.

Included code to integrate with the database. It is commented out for now.
