+++ src/pages/nav/addProblems.js
@@ -1,15 +1,13 @@
-/**
- * Adding a problem page
- * This page allows us to add a new problem to the database
- * given a title, description, example inputs, example outputs,
- * along with a solution/hint
- *
- * This is temporary and will be avilable only to the admin
- * once we adminster users and roles
- */
-
 import React, { useState } from 'react'
 import { Box, Stack, Button, TextField } from '@mui/material'
+import TestCase from '../../components/add-problems/TestCase'
+import {
+  handleChange,
+  handleTestCaseChange,
+  addTestCase,
+  removeTestCase,
+  handleSubmit,
+} from '../../components/add-problems/ICPCFormHandlers'
 
 const AddProblems = () => {
   const [formData, setFormData] = useState({
@@ -23,72 +21,19 @@ const AddProblems = () => {
     testCases: [{ input: '', output: '' }],
   })
 
-  const handleChange = (e) => {
-    const { name, value } = e.target
-    setFormData({
-      ...formData,
-      [name]: value,
-    })
-  }
-
-  const handleTestCaseChange = (index, e) => {
-    const { name, value } = e.target
-    const testCases = [...formData.testCases]
-    testCases[index][name] = value
-    setFormData({
-      ...formData,
-      testCases,
-    })
-  }
-
-  const addTestCase = () => {
-    setFormData({
-      ...formData,
-      testCases: [...formData.testCases, { input: '', output: '' }],
-    })
-  }
-
-  const removeTestCase = (index) => {
-    const testCases = formData.testCases.filter((_, idx) => idx !== index)
-    setFormData({
-      ...formData,
-      testCases,
-    })
-  }
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-    const payload = { ...formData }
-    console.log(JSON.stringify(payload, null, 2)) // or send the payload to a server
-
-    alert('Input submitted!')
-
-    // Clear the form
-    setFormData({
-      title: '',
-      timeLimit: '',
-      memoryLimit: '',
-      problemDescription: '',
-      inputDescription: '',
-      outputDescription: '',
-      videoLink: '',
-      testCases: [{ input: '', output: '' }],
-    })
-
-    // Reload the page
-    window.location.reload()
-  }
-
   return (
     <Box component="section" display={'flex'} padding={3} width={'95%'}>
-      <form onSubmit={handleSubmit} style={{ width: '100%' }}>
+      <form
+        onSubmit={(e) => handleSubmit(e, formData, setFormData)}
+        style={{ width: '100%' }}
+      >
         <Stack spacing={2}>
           <label>Title: </label>
           <TextField
             type="text"
             name="title"
             value={formData.title}
-            onChange={handleChange}
+            onChange={(e) => handleChange(e, formData, setFormData)}
             required
             label="Please input a title"
             variant="outlined"
@@ -97,7 +42,7 @@ const AddProblems = () => {
           <TextField
             name="problemDescription"
             value={formData.problemDescription}
-            onChange={handleChange}
+            onChange={(e) => handleChange(e, formData, setFormData)}
             required
             label="Please input a problem description"
             variant="outlined"
@@ -107,7 +52,7 @@ const AddProblems = () => {
             type="text"
             name="timeLimit"
             value={formData.timeLimit}
-            onChange={handleChange}
+            onChange={(e) => handleChange(e, formData, setFormData)}
             label="Please input a time limit"
             variant="outlined"
           />
@@ -116,7 +61,7 @@ const AddProblems = () => {
             type="text"
             name="memoryLimit"
             value={formData.memoryLimit}
-            onChange={handleChange}
+            onChange={(e) => handleChange(e, formData, setFormData)}
             label="Please input a memory limit"
             variant="outlined"
           />
@@ -124,16 +69,16 @@ const AddProblems = () => {
           <TextField
             name="inputDescription"
             value={formData.inputDescription}
-            onChange={handleChange}
-            label="Please input a input description"
+            onChange={(e) => handleChange(e, formData, setFormData)}
+            label="Please input an input description"
             variant="outlined"
           />
           <label>Output Description: </label>
           <TextField
             name="outputDescription"
             value={formData.outputDescription}
-            onChange={handleChange}
-            label="Please input a output description"
+            onChange={(e) => handleChange(e, formData, setFormData)}
+            label="Please input an output description"
             variant="outlined"
           />
           <label>Video: </label>
@@ -141,13 +86,13 @@ const AddProblems = () => {
             type="text"
             name="videoLink"
             value={formData.videoLink}
-            onChange={handleChange}
+            onChange={(e) => handleChange(e, formData, setFormData)}
             label="Please input a video link"
             variant="outlined"
           />
           <div>
             <label>Test Cases: </label>
-            <Button variant="contained" type="button" onClick={addTestCase}>
+            <Button variant="contained" type="button" onClick={() => addTestCase(formData, setFormData)}>
               Add Test Case
             </Button>
             <Stack
@@ -157,39 +102,13 @@ const AddProblems = () => {
               spacing={3}
             >
               {formData.testCases.map((testCase, idx) => (
-                <Stack
+                <TestCase
                   key={idx}
-                  direction="row"
-                  justifyContent="center"
-                  alignItems="center"
-                  spacing={2}
-                >
-                  <label>Input: </label>
-                  <TextField
-                    type="text"
-                    name="input"
-                    value={testCase.input}
-                    onChange={(e) => handleTestCaseChange(idx, e)}
-                    label=""
-                    variant="standard"
-                  />
-                  <label>Output: </label>
-                  <TextField
-                    type="text"
-                    name="output"
-                    value={testCase.output}
-                    onChange={(e) => handleTestCaseChange(idx, e)}
-                    label=""
-                    variant="standard"
-                  />
-                  <Button
-                    variant="contained"
-                    type="button"
-                    onClick={() => removeTestCase(idx)}
-                  >
-                    Remove
-                  </Button>
-                </Stack>
+                  index={idx}
+                  testCase={testCase}
+                  handleTestCaseChange={(index, e) => handleTestCaseChange(index, e, formData, setFormData)}
+                  removeTestCase={() => removeTestCase(idx, formData, setFormData)}
+                />
               ))}
             </Stack>
           </div>
